Created test structure, /api/v1/new positive test
