+++ Cargo.toml
@@ -4,6 +4,13 @@ version = "0.1.0"
 authors = ["cascalheira"]
 edition = "2018"
 
+[lib]
+path = "src/lib.rs"
+
+[[bin]]
+path = "src/main.rs"
+name = "authenticator"
+
 [dependencies]
 actix-web = { version = "3", features=["rustls"] }
 actix = "0.10"
@@ -23,3 +30,6 @@ sled = "0.34.6"
 sled-extensions = {version = "0.2.0", features = ["bincode"]}
 chrono = "0.4"
 once_cell = "1.8.0"
+
+[dev-dependencies]
+actix-rt = "1.1.1"
